requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
flake8==7.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
